style: 파일 디렉토리 구조를 변경한다
