+++ index.html
@@ -4,6 +4,7 @@
     <meta charset="UTF-8" />
     <title>🏎️ 자동차 경주 게임</title>
     <link rel="stylesheet" href="./src/css/index.css" />
+    <script type="module" src="./src/js/index.js"></script>
   </head>
   <body>
     <div id="app">
@@ -50,36 +51,7 @@
         </div>
       </div>
       <div class="d-flex justify-center mt-5">
-        <section class="game-progress-container mt-4">
-          <!--
-          <div class="d-flex">
-            <div>
-              <div class="car-player mr-2">EAST</div>
-              <div class="forward-icon mt-2">⬇️️</div>
-              <div class="forward-icon mt-2">⬇️️</div>
-            </div>
-            <div>
-              <div class="car-player mr-2">WEST</div>
-              <div class="forward-icon mt-2">⬇️️</div>
-            </div>
-            <div>
-              <div class="car-player mr-2">SOUTH</div>
-              <div class="d-flex justify-center mt-4">
-                <div class="relative spinner-container">
-                  <span class="material spinner"></span>
-                </div>
-              </div>
-            </div>
-            <div>
-              <div class="car-player mr-2">NORTH</div>
-              <div class="d-flex justify-center mt-4">
-                <div class="relative spinner-container">
-                  <span class="material spinner"></span>
-                </div>
-              </div>
-            </div>
-          </div>
-        --></section>
+        <section class="game-progress-container mt-4 d-flex"></section>
       </div>
       <div class="d-flex justify-center mt-5">
         <section class="game-result-container" hidden>
