--- src/js/util/constants.js
@@ -1,27 +1,0 @@
-const SELECTOR = {
-  TITLE: {
-    CONTAINER: ".title-container",
-  },
-
-  CARNAME: {
-    CONTAINER: ".carname-input-container",
-    INPUT: ".carname-input-container__input",
-    BUTTON: ".carname-input-container__button",
-  },
-
-  LAP: {
-    CONTAINER: ".lap-input-container",
-    INPUT: ".lap-input-container__input",
-    BUTTON: ".lap-input-container__button",
-  },
-
-  GAME_PROGRESS: {
-    CONTAINER: ".game-progress-container",
-  },
-
-  GAME_RESULT: {
-    CONTAINER: ".game-result-container",
-  },
-};
-
-export { SELECTOR };
